rename package and move to 'client' dir
